split tests in separate file
